RZ Pack
Moved two lessons out of the Beta pack into a new pack, made a new pack
for upcoming training.

Not sure where I should put the name of the content pack, I would name
it:

“Radio Zamaneh”
“Beta StoryPaths for community participation.”

I have put the cover images into the root directory of the folder “rz”
in the assets folder.