Mobile Photo v4 Updates
New lessons added to pack:

* Know Your Mobile
* Shoot on an Angle
* Shoot on an Angle Activity
* Control Light
* Control Light Activity

The order has been adjusted, “Know Your Mobile” is now the first
lesson. Shoot an Angle and Control Light have been added to the end.

This is the last major update to the pack, once the curriculum team can
review this, they will make another small round of tweaks/adjustments
based on input from dev team on what is possible with the milestone
cards, and some adjustments to the multimedia assets as well.

Specific Details: '

Updated the content index to account for the changes.
Added new folders and new yaml for all new paths.
Adjusted library files to reflect new names of folders.