Mobile Photo V5 (Final Content Updates)
Updated yaml and assets folders with version 5 changes, and new content
from the curriculum team. From this point forward, no significant
changes will be happening.

As a proactive move to see how possible a table of contents may be I
went ahead and added a Table of Contents Path. It’s the  "toc.yaml" in
the "0_toc" folder as a way to test out the ability to open a single
path when opening a content pack. Please advise on your
time/availability to see if this is possible.