final edits to v1 of video_stories and process_stories content packs
These edits are primarily formatting, and removing some legacy content
(such as EvaluationCard). No further edits are expected until after the
packs are released and can be tested live.

This leaves photo_essay_stories as the final content pack to be
completed for the next build. That content pack will be completed
tomorrow (12/9).