fixes to template libraries
This fix involved the addition of dependencies, as well as fixing some
incorrectly named paths, which meant the user could not choose a path,
or caused the “more info” button to not work.

I still need to commit a “more info” milestone card for t_process. I
can do that this evening.